ai-service/requirements.txt
flask==3.1.1
google-generativeai==0.8.3
python-dotenv==1.1.0
